Organized test files under action & func sub-dirs
