+++ test/unit/CMakeLists.txt
@@ -1,8 +1,8 @@
 cmake_minimum_required(VERSION 3.6)
 
 add_executable(polyC_unit_tests testMain.cpp
-        src/SingleImplementationActionTests.cpp src/SingleImplementationFuncTests.cpp
-        src/MultiImplementationActionTests.cpp src/MultiImplementationFuncTests.cpp)
+        src/action/SingleImplementationActionTests.cpp src/func/SingleImplementationFuncTests.cpp
+        src/action/MultiImplementationActionTests.cpp src/func/MultiImplementationFuncTests.cpp)
 
 target_include_directories(polyC_unit_tests PRIVATE include)
